Worked on homework

+++ Homework_Assignments/Homework_2/CDA_Homework_2/Electrical_Energy/Electrical_Energy.cpp
@@ -7,7 +7,7 @@
 	#include <iostream>
 	#include <vector>
 	using  namespace  std;
-	int counter = 0;
+	long long counter = 0;
 	void merge(long long* arr, long long* helper, long long start, long long mid, long long end)
 	{
